feat: 218 propagate baggage and Stanza specific headers to outgoing request
* create withStanzaHeaders function
* clean up index file from Stanza Cloudflare SDK
* fix tsconfig.json from Stanza Cloudflare SDK to not include dom libs
