+++ packages/sdk-cloudflare/src/headerSetter.ts
@@ -1,0 +1,10 @@
+import type { TextMapSetter } from '@opentelemetry/api';
+
+export const headerSetter: TextMapSetter<Headers> = {
+  set(carrier, key, value) {
+    if (carrier == null) {
+      return;
+    }
+    carrier.set(key, value);
+  },
+};
